requirements.txt
langchain==1.0.5
langchain_community==0.4.1y
langchain-google-genai==3.0.3
python-dotenv
chromadb == 1.3.4
langchain-classic